refactor: add .js extension to imports
