+++ src/ts/index.ts
@@ -1,4 +1,4 @@
-export { default as secure2PC } from "./secure2PC";
-export { default as BufferedIO } from "./BufferedIO";
-export { default as BufferQueue } from "./BufferQueue";
+export { default as secure2PC } from "./secure2PC.js";
+export { default as BufferedIO } from "./BufferedIO.js";
+export { default as BufferQueue } from "./BufferQueue.js";
 export { type IO } from "./types";
