+++ src/ts/secure2PC.ts
@@ -1,7 +1,7 @@
 import { EventEmitter } from "ee-typed";
 import type { IO } from "./types";
-import workerSrc from "./workerSrc";
-import nodeSecure2PC from "./nodeSecure2PC";
+import workerSrc from "./workerSrc.js";
+import nodeSecure2PC from "./nodeSecure2PC.js";
 
 export type Secure2PC = typeof secure2PC;
